+++ src/main/java/org/jsoup/parser/StreamParser.java
@@ -7,7 +7,7 @@ import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 import org.jsoup.select.Evaluator;
 import org.jsoup.select.NodeVisitor;
-import org.jsoup.select.QueryParser;
+import org.jsoup.select.Selector;
 import org.jspecify.annotations.Nullable;
 
 import java.io.Closeable;
@@ -209,7 +209,7 @@ public class StreamParser implements Closeable {
      @see #selectFirst(Evaluator)
      */
     public @Nullable Element selectFirst(String query) throws IOException {
-        return selectFirst(QueryParser.parse(query));
+        return selectFirst(Selector.evaluatorOf(query));
     }
 
     /**
@@ -236,7 +236,7 @@ public class StreamParser implements Closeable {
      @param eval the {@link org.jsoup.select.Selector} evaluator.
      @return the first matching {@link Element}, or {@code null} if there's no match
      @throws IOException if an I/O error occurs
-     @see QueryParser#parse(String)
+     @see Selector#evaluatorOf(String css)
      */
     public @Nullable Element selectFirst(Evaluator eval) throws IOException {
         final Document doc = document();
@@ -257,7 +257,7 @@ public class StreamParser implements Closeable {
      @see #selectNext(Evaluator)
      */
     public @Nullable Element selectNext(String query) throws IOException {
-        return selectNext(QueryParser.parse(query));
+        return selectNext(Selector.evaluatorOf(query));
     }
 
     /**
@@ -284,7 +284,7 @@ public class StreamParser implements Closeable {
      @param eval the {@link org.jsoup.select.Selector} evaluator.
      @return the next matching {@link Element}, or {@code null} if there's no match
      @throws IOException if an I/O error occurs
-     @see QueryParser#parse(String)
+     @see Selector#evaluatorOf(String css)
      */
     public @Nullable Element selectNext(Evaluator eval) throws IOException {
         try {