+++ src/main/java/org/jsoup/select/Elements.java
@@ -525,7 +525,7 @@ public class Elements extends ArrayList<Element> {
      * @return true if at least one element in the list matches the query.
      */
     public boolean is(String query) {
-        Evaluator eval = QueryParser.parse(query);
+        Evaluator eval = Selector.evaluatorOf(query);
         for (Element e : this) {
             if (e.is(eval))
                 return true;
@@ -603,7 +603,7 @@ public class Elements extends ArrayList<Element> {
 
     private Elements siblings(@Nullable String query, boolean next, boolean all) {
         Elements els = new Elements();
-        Evaluator eval = query != null? QueryParser.parse(query) : null;
+        Evaluator eval = query != null? Selector.evaluatorOf(query) : null;
         for (Element e : this) {
             do {
                 Element sib = next ? e.nextElementSibling() : e.previousElementSibling();