+++ src/main/java/org/jsoup/select/Evaluator.java
@@ -23,7 +23,7 @@ import static org.jsoup.internal.StringUtil.normaliseWhitespace;
 
 /**
  An Evaluator tests if an element meets the selector's requirements. Obtain an evaluator for a given CSS selector
- with {@link QueryParser#parse}. If you are executing the same selector on many elements (or documents), it
+ with {@link Selector#evaluatorOf(String css)}. If you are executing the same selector on many elements (or documents), it
  can be more efficient to compile and reuse an Evaluator than to reparse the selector on each invocation of select().
  <p>Evaluators are thread-safe and may be used concurrently across multiple documents.</p>
  */